Use registry to check Atime support on Windows
