+++ cmd/disk-cache-check-support_other.go
@@ -1,0 +1,56 @@
+// +build !windows
+
+/*
+ * MinIO Cloud Storage, (C) 2019-2020 MinIO, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cmd
+
+import (
+	"errors"
+	"io"
+	"io/ioutil"
+	"os"
+	"time"
+
+	"github.com/djherbis/atime"
+)
+
+// Return error if Atime is disabled on the O/S
+func checkAtimeSupport(dir string) (err error) {
+	file, err := ioutil.TempFile(dir, "prefix")
+	if err != nil {
+		return
+	}
+	defer os.Remove(file.Name())
+	defer file.Close()
+	finfo1, err := os.Stat(file.Name())
+	if err != nil {
+		return
+	}
+	// add a sleep to ensure atime change is detected
+	time.Sleep(10 * time.Millisecond)
+
+	if _, err = io.Copy(ioutil.Discard, file); err != nil {
+		return
+	}
+
+	finfo2, err := os.Stat(file.Name())
+
+	if atime.Get(finfo2).Equal(atime.Get(finfo1)) {
+		return errors.New("Atime not supported")
+	}
+	return
+}
